\name{CRImage-internal}
\alias{findSlices}
\alias{classifyStructures}
\alias{numberOfNeighbors}
\alias{parseFinalScan}
\alias{paintCells}
\alias{kernelSmoother}
\alias{classificationAperio}
\alias{searchStructures}
\alias{segmentCytoplasma}
\alias{writeDensityImage}
\alias{getCorrected}
\alias{localOtsuThreshold}
\alias{localORThreshold}
\alias{Phansalkar_threshold}
\alias{imageCompression}
\alias{determineCellularity}
\alias{imageCompression}
\alias{segmentStructures}

\title{Internal CRImage Functions}
\description{
 Internal CRImage functions
}

\details{
These are internal functions, which is not to be called by the user.
}
\keyword{ internal }